feat: start frontend building

+++ server/controllers/product.controller.js
@@ -47,12 +47,14 @@ const updateProduct = async (req, res) => {
 const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if(!mongoose.Types.ObjectId.isValid(id)) return res.status(404).json({ success: false, message: 'ID inválido.' })
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: 'Produto deletado.' });
   } catch (error) {
     console.error('> Error while trying to delete project');
-    res.status(200).json({ success: false, message: 'Produto não encontrado.' });
+    res.status(500).json({ success: false, message: 'Erro interno no servidor.' });
   };
 };
